Fix potential bugs reported by compiler

+++ src/rydeen.h
@@ -6,6 +6,6 @@
 #define debug(...)
 #endif
 
-#define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
+#define ARRAY_SIZE(arr) (int)(sizeof(arr) / sizeof((arr)[0]))
 
 extern struct ev_loop *loop;
